Fixing a bug, a 508 complaint, and optimizing FormShow with denormalizr
